Clarify what portable_audio_player.folder_depth means.
> > Assuming that a device is not listed in the HAL fdi
> > entry as having a limited depth, most software would probably use
> > Artist/Album/Track when appropriate.  If you think this is not
> > sufficient to support the W800i, let's certainly keep discussing it.
>
> That would indeed seem to work fine for the W800i, it just seems open
> to more problems than the previous method of allowing an exact
> specification. While I agree that software would probably use
> Artist/Album/Track it might use Artist - Album/Track which doesn't
> work.

The real problem here is that your device requires a minimum (exact,
actually) folder depth, right?  I think the best solution is to make
the folder_depth field not be a max limit, but to be the exact depth
required by a device.

I've attached a patch that just clarifies the role of the
portable_audio_device.folder_depth key.  I think this is a pretty
simple way to solve this problem, and won't be too hard for app
developers to implement.