Close memory leaks from not freed DBusError in libhal and libhal-storage
Added a macro to libhal to free a DBusError (including a check if the error
was really set and a debug message if this is not the case) and closed several
memory leaks in the libhal and libhal-storage library.

Did some little cleanups in the libraries and fixed wrong displayed/returned
DBusError in libhal_ctx_shutdown().