Close memory leaks from not freed DBusError in tools src dir
Closed several memoryleaks related to not freed DBusError objects
in *.c files in the tools dir by using LIBHAL_FREE_DBUS_ERROR macro.
This include also a cleanup in hal-storage-mount.c and a fix for lshal
to avoid start monitoring device list if we can't set the watch to
all device properties.