make libhal link against libintl for NetBSD (bug #6471)
libhal uses i18n functions such as bindtextdomain.  These functions are not
always provided by the standard C library (as happens in Linux).  For example,
NetBSD has the i18n functions in libintl.

Therefore, libhal should be linked against the appropriate library that provides
the missing functions.  This library is available in the INTLLIBS variable
thanks to the AM_GLIB_GNU_GETTEXT variable used in configure.

If libhal is not linked to libintl, any program using libhal fails to build
because the missing functions cannot be found.  This happens while trying to
build the tools directory.