fix up how we handle non-partitioned volumes
We used to only creat a fakevolume if a mountable file system was
detected. This is wrong as you can have LUKS, RAID or LVM on the main
block device and we should detect this just like we do on
partitions. This patch fixes that.

Second, assume a piece of media is all zero'ed out. Then, on
insertion, we don't find anything on the main block device and neither
does the kernel so no child block devices are created. Hence, HAL
won't spit out any event when such media is inserted. And that, so
far, is correct as we don't detect any useful volumes. However, it
would be useful in e.g. a disk utility program to know if media is
actually available or not. This patch introduces a new property called
storage.removable.media_available that is TRUE if, and only if, we have
detected media on the storage device. Also expose this new property
in libhal-storage.