reduce useless changes on APM battery.remaining_time
Changed update property battery.remaining_time on APM to avoid change the
value every 2 seconds. APM calculate with each poll a new value for
remaining battery time, which could result in high dbus load if a application
listen for the event and check do something over the bus. On the other side
the remaining time from the APM interface is not trustworthy on the most
machines if you poll to often, increase the poll intervall (and emit a new
value to HAL) should smooth the time curve and reduce back and forth jumping
values.

Now the value is updated ever 30 seconds, this should preduce the needless
events.