properly convert mAh to mWh rather than uWh
We should convert the mAh reading (battery.reporting.rate) into mWh
(battery.charge_level.rate) by multiplying it by the voltage (in Volts),
but we are multiplying it by the voltage in mV, and so the rate information
is a factor of 3 too large. This *really* breaks calculating the time
for multibattery laptops.