fix message when we try to set a brightness above range
Setting the brightness above range on the command line gives:

[hughsie@hughsie-laptop hal]$ dbus-send --system --print-reply
--dest=org.freedesktop.Hal /org/freedesktop/Hal/devices/acpi_lcd
org.freedesktop.Hal.Device.LaptopPanel.SetBrightness int32:8
Error org.freedesktop.Hal.Device.LaptopPanel.Invalid: Brightness has to
be between 0 and 8!

So the message is telling us to set a value between 0-8, which we really
are trying to do! The correct message would be to tell us the range was
0->7.