fix ACPI acpid/proc configure options
Fix for fd.o bug 7278. From Bugzilla: "... HAL contains a flawed configure.in
that disables acpi-proc support when --enable-acpi-proc is passed as an
argument. --disable-acpi-proc correctly disables it, and not specifying
anything correctly enables it. This is because the current code does not
check the $enableval as it should.  The acpi-acpid enable flag is similarly
broken. This causes acpi support to be disabled for distributions that
attempt to enable it via --enable-acpi-*.  I've attached a patch for the
configure.in from CVS HEAD which should fix the problem, allowing
enable-acpi-* and disable-acpi-* to be correctly detected and enabled by
default."

NOTE: You maybe need to check and adapt your current configure options for
      the ACPI settings. The behavior maybe differ from the current.