add new partition probing code and adapt hal code to use it
Added a new static library in partutil/ that serves two purposes

 1. Probe for partition tables and the entries in them. This is
    done without any new dependencies and is in general efficient.
    Right now we understand Master Boot Record (and Extended MBR),
    GUID Partitioning Tables and Apple Partition Map. The interface
    is generic enough to add support for other disk labels too
    should the need arise.

 2. Create partition tables, add partition table entries, alter
    partition table entries, delete partition table entries. This
    is done using libparted to ensure maximal robustness and
    correctness.

Beware that libparted is kinda silly in a few ways, it probes all
drives (even when asked to only look at one drive - the audicity!),
and we have to poke at the internals to get/set vital raw
properties. As such, we only officially support libparted 1.7.1 at
this point. If you use anything else but that, you are on your own.
We should start a dialouge with the libparted guys to make them
export official API suitable for our needs.

configure now takes a new --enable-parted and by default (for
the upcoming 0.5.8 release), this is set to "no". As such, the
upcoming 0.5.8 release will not depend on libparted.

Also we now use this infrastructure in probe-volume and probe-storage
to export the following new properties

 storage.partitioning_scheme (string)

on storage devices and

 volume.partition.scheme (string)
 volume.partition.type (string)
 volume.partition.label (string)
 volume.partition.uuid (string)
 volume.partition.flags (strlist)

on volumes. These new properties are also now exported through
libhal-storage.

Also moves some more code over to using LibHalChangeSet for nicer
and much more efficient handling.