enable looking for data track on optical discs
This fixes two bugs; a) when a mixed disc got data as the last track
(the usual case for bonus CD's); and b) picking the label from the
last session. This was actually disabled by Kay some time ago with a
note that a few drives have broken firmware. We need to blacklist
those.