don't add non-existing partition entries for MS-DOS disk labels
I think the issue is that the users who run into this have broken
partition extended tables, e.g. we error out in the function
part_table_probe_msdos_extended() and thus return NULL for the
extended partition table. We then stuff a NULL pointer into the
entries list and that's causing the crash on freeing the partition
table or looking through entries. This patch should fix at least the
crashes.