some memory optimization work
There's an interesting behavior on startup. On my machine we allocate
around 300 HotplugEvent objects (each of 3604 bytes) on startup and
put them on the queue. When we're done processing an event it's freed.
This leads to interesting memory fragmentation problems. Some
experimental code I've just put in allocates the first 1000 in a fixed
array. Seems to save 200-300kb writable memory.

Also move some bits of HalDevice into HalDevicePrivate. Fix up the
rest of the code not to rely on the ->udi member as it's now only
available through the existing hal_device_get_uid() method.