don't call dbus_connection_close() for some of our tools
lshal and hal-device should not call dbus_connection_close() as they do
not have a private connection to the bus.

This is not critical. However, with D-Bus 0.94+ it will spawn ugly error
messages.  Note: It is safe to apply this patch for HAL when using older
versions of D-Bus, too.