ignore the special ACPI 'Ones' value for rate
As discovered in : http://bugzilla.gnome.org/show_bug.cgi?id=348201
ACPI gives out the special 'Ones' value for rate when it's unable to
calculate the true rate.
We should set the rate zero, and wait for the BIOS to stabilise for the
few BIOS's that give 0xffff as the very first entry after an AC state
change.
This patch has been tested to work by Pierre Ossman.