fix copy_property for merge in FDI files
This patch fixed a bug with copy_property in FDI files if you try to
copy a property from an other device. If you try:

  <merge key="baz" type="copy_property">@info.parent:info.vendor</merge>

The key get copied to the info.parent device instead of the currently
processed device. The bug was introduced with commit 1ffd7638 by reuse a
variable name for a temporary local variable.