replace usage of g_assert() with error handling
This removes the only one g_assert() within the complete daemon from
blockdev.c and replace this with code to check also the tdl instead of
only gdl and go to the (already existing) error case if HAL don't get
the device  searching for and simply print a warning and error mesage
and don't add the device to the device tree. In worst case we have
just have one device less instead of dead HAL.

Without the fix HAL die without any message sometimes on boot. See:
https://bugzilla.novell.com/show_bug.cgi?id=217563 or
https://bugs.freedesktop.org/show_bug.cgi?id=8210