add /sys/class/backlight/* interface support
If available, a device /org/freedesktop/Hal/devices/computer_backlight
is created. The change adds a class device and adjusts the scripts for
the methods GetBrightness and SetBrightness on the interface
org.freedesktop.Hal.Device.LaptopPanel. As soon as all kernel drivers are
all converted to the new interface, we can drop the vendor specific code in
acpi.c. In case a driver supports both, the old /proc interface and the
new /sysfs one, the code in acpi.c immediately returns so that the /sysfs
interface is preferred.