Make hotplugging non-recursive.
This adds a notify callback for when hald_runner has finished calling a callout.
For linux, we install hotplug_event_process_queue as the notify, and modify hotplug_event_process_queue to iterate the queue. All recrursive calls to hotplug_event_process_queue are removed. This results in massive stack and a not inconsiquential heap savings.