Remove linux.sysfs_path_device key
linux.sysfs_path_device is no longer needed internally in HAL.
Removing without going through deprectation process as this key was documen
as only used internally and not guaranteed to exist, so noone should be usi
it.
Also fixes problems with mmc and scsi devices getting incorrectly parented in
the device tree.