Fix to guarantee that udi get generated before process FDI
This fix change the existing workflow of HAL to guarantee that the UDI of
a device get generated before start processing the FDI files.

The problem is: in the most cases the udi of a device get generated fist
after the policy and information FDI files get processed. Because of this
you can't use FDI rules to match a special device udi (the udi of the
device is something like '/org/freedesktop/Hal/devices/temp/100' until a
correct udi get generated) e.g. to spawn a child device because the rule
get never proccessed again for this device.

To fix this I moved the generation of the udi before the calls to process
the policy and information fdi-files. The mean also to move some
hal_util_callout_device_add() calls (to process the FDI files after the
call).

This was tested for the linux backend and showed no adverse effects. The
FreeBSD backend was also changed, but could not get tested.