fix pmu laptop screen on new kernels
New kernels export the PMU backlight as a kernel backlight class, and
GET_BACKLIGHT and SET_BACKLIGHT ioctls fail. This means that HAL fails
to set the brightness on PPC hardware.

We should check for the existence of the backlight device (like in
acpi.c) before we add the PMU backlight to avoid having two backlight
devices, one of which that does not work.

The patch also gives the generic backlight class a name, so it does not
show up as unknown device.