add optional support for ConsoleKit
You have to enable this with --enable-console-kit. Right now the way
it works is that service (e.g. methods on interfaces that is not
org.freedesktop.Hal.Device) is denied to inactive sessions. Presently
the main hald process denies this; going forward we will probably move
this to specific method handlers (program being run such as Suspend()
etc.) and add-on's.

There's an easy to fix bug in pam_console on Fedora Rawhide for this
to work out of the box on Rawhide, see

 https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=227462

for details and how to fix it on your local system.

Also, instead of asking the bus daemon for the UNIX user id of the
caller we now cache it per connection. So there's a small optimization
insofar that we save a roundtrip here and there.