fix detection of max. available omnibook brightness levels
This patch fix the the number of max. available brightness levels for the
proc omnibook interface.

The reason for this change:
* there exist two different versions of the omnibook kernel module (the older
version is from [1] and the new one from project [2]).
* different hardware supported by this module have different max levels.
(older machines have 11 (0-10) and newer hardware support 8 levels (0-7))

How this patch handle this:
* for the newer module version the max. level can be parsed
from /proc/omnibook/lcd
* for the older version we assume now always max. 11 levels, since this module
only support level 0-10 for brightness.

[1] http://sourceforge.net/projects/omke/
[2] http://sourceforge.net/projects/omnibook/