added new section for deprecated properties to spec
Added a new section to spec to collect information about deprecated
properties. The new table provide info about the deprecated key, the
key which replace the deprecated one (if there is on), the earliest
possible date for complete deletion of the property and a column for
note/explanation.

Added info for info.bus, *.physical_device and the deprecated
power_management.can_suspend_to_* keys to the new section.