two build fixes
It seems that in order to detect libsmbios correctly the AC_CHECK_LIB
needs to be wrapped with AC_LANG_PUSH([C++]) and AC_LANG_POP([C++]).

Also if --disable-gtk-doc is used then all the PKG_CHECK_MODULES()
fail.  It seems that pkg-config stuff is only initialized in the
GTK_DOC_CHECK macro.