fix .hal-mtab mountpoint handling
This fixes problems with the mountpoint handling in .hal-mtab and non ASCII
chars as e.g. in languages like Russian. This patch remove the usage of
g_strescape() to escape the mountpoint for .hal-mtab.

Tested this with the string from [1] (also reported as fd.o bug #9921) and
normal ASCII label names (also with space characters in the mountpoint) and
there was no problem and both cases worked as expected for mount and umount
now.

[1] https://bugzilla.novell.com/show_bug.cgi?id=242466