remove hal-spec.html from the tarball
On Tue, 2007-02-20 at 01:11 -0500, David Zeuthen wrote:
> as it rendered the commit logs virtually unreadable. Downside is that
> it's a bit harder to look at the current spec but we ought to solve
> that
> in other ways.
>
> Btw, feel free to commit a fix for this. Thanks :-)

Well, we have two options:

1. Force anybody that does make distcheck to use --enable-docbook-docs
so that the html file gets created.
2. Remove the html from the tarball

I personally think that the hal spec should be a static html file on the
hal fd.o page (and therefore removed from the tarball), or created using
a post-commit hook as Rob suggested.

Richard.