use inotify to watch fdi directories
When we see activity in the fdi file directories we call a method to
invalidate the cache. The next time fdi files are needed, the cache
will be regenerated. Currently it takes around 93 ms on my Core Duo
2GHz to regenerate the cache (it's roughly 170kb).

Solaris and FreeBSD backends should implement a similar mechanism and
call osspec_fdi_cache_invalid() when changes are seen.