fix acpi detection for very new kernels
Apply patch from S.Çağlar Onur <caglar@pardus.org.tr> to make
HAL detect the presence of ACPI not from the info file, but
the acpi directory as the info file is not in sysfs.
Slightly tweaked by me to remove a warning.

You will need to use HAL 0.5.9 with >2.6.21 if you are not using the
configuration option CONFIG_ACPI_PROCFS.