added new fdi match attribute 'contains_not' for strings and strlist
Added a new attribute for FDI file match tag for string lists (strlist)
and strings. The new attribute named contains_not and allow you to match
if a given string is not contained in the string list/strins as in such
cases for example:

 <match key="info.callouts" contains_not="foobar">
  <append key="info.callouts" type="strlist">foobar</append>
 </match>

This allow you to add e.g. callouts to the list without to double entries
and allow some kind of if/else clause in FDI files