validate error D-Bus names returned on stderror
Attached a patch to validate the error name and workaround the problem in HAL.
Because there is no funtion in the D-Bus libraries to check if a interface or
error name is valid, I adapted the D-Bus internal funtion for util.h. If the
check fail and the error name is not valid the patch set the error name to:

        org.freedesktop.Hal.Device.UnknownError

The invalid error name get prepend to the error message to avoid lost information.

Maybe we should change the error name to:

        org.freedesktop.Hal.Device.MalformedError