handle system bus restarts
the following patch is preventing hald to exit when system bus is
restarted (I won't troll^W comment on whether it should survive or not),
just like previous release of hal.

As a side note, I've noticed upstream dbus (as well as Fedora package)
is shipping with fatal_warning_on_check (I'll let interested people to
refer to flamewa^Wdiscussion about that last november on dbus mailing
list), in the hope of getting dbus based applications to fix their
incorrect usage of dbus API, which is why we are seeing hald exiting a
lot. Debian / Ubuntu (as now Mandriva) are shipping with those checks
disabled, so this might explain why not everybody sees hald exiting more
than it should.