change usb*.version_bcd/usb*.speed_bcd to new double keys
This adds two new keys which store the version and the speed of a USB
device as double values and not in BCD. This looks better in lshal and
should also be easier to handle by applications which consume this
information.

Set the old usb*.*_bcd properties in the spec to deprecated.