add named mandatory locking API
This provides a mechanism for clients to acquire (and release) named
locks on device objects. More than one client can hold a named lock on
a device, this is by design. If a client disconnects from the bus, it
locks will be removed.

The rationale for this new API is to allow clients to lock out other
clients of service on given device objects; this is applicable for

 - installers (inhibit mount requests)

 - partitioning/formatting tools (inhibit mount requests)

 - multi-seat power management (inhibit suspend/hibernate etc. requests
   from idle (e.g. the user is idle) active (e.g. it is in the foreground
   on some seat) sessions that wants to suspend due to user preferences.)

The next series of patches will make the various mechanisms use this
API by defining well-known locks.