provide textual interface description for USB devices
Many USB devices provide textual descriptions for their interfaces and
configurations. Currently HAL ignores them. For example, I have a Nokia 6680
phone which has two OBEX interfaces that provide different services, but the
only way to tell that is to read interface descriptions:

    Interface Descriptor:
      bLength                 9
      bDescriptorType         4
      bInterfaceNumber        4
      bAlternateSetting       0
      bNumEndpoints           0
      bInterfaceClass         2 Communications
      bInterfaceSubClass     11 OBEX
      bInterfaceProtocol      0
      iInterface              5 SYNCML-SYNC
      CDC Header:
        bcdCDC               1.10
      CDC OBEX:
        bcdVersion           1.00
      CDC Union:
        bMasterInterface        4
        bSlaveInterface         5
....
    Interface Descriptor:
      bLength                 9
      bDescriptorType         4
      bInterfaceNumber        6
      bAlternateSetting       0
      bNumEndpoints           0
      bInterfaceClass         2 Communications
      bInterfaceSubClass     11 OBEX
      bInterfaceProtocol      0
      iInterface              6 PC Suite Services
      CDC Header:
        bcdCDC               1.10
      CDC OBEX:
        bcdVersion           1.00
      CDC Union:
        bMasterInterface        6
        bSlaveInterface         7

From reading the source code, it seems like an easy to fix thing to me. At least
on Linux these descriptions are available as 'interface' and 'configuration'
files in sysfs directories for the USB interfaces and devices respectively. The
specification should be corrected also of course. If you want, I can write the
patch.