work around cdrom devices giving invalid TOC info
A friend of mine has a silly dvd drive, which gives a wrong offset for
certain discs[0] causing hal to fail when trying to detect a file
system.

Attached patches (one for master, one for 0.5.8) work around this be
retrying detection at offset 0 when it fails at the reported offset.

  Sjoerd

0: All dvd's we tried had an offset of 0 for their (only) session, but the drive
indicated an offset of -150.