allow hibernate, suspend and CPUFreq for the root user independent of desktop-console
This changes the logic how privileges for hibernate, suspend and cpufreq
are determined. Now, the root user is granted to do hibernate, suspend and
all CPUFreq methods with no regard to the desktop-console privilege. In the
past, the application/script checking if the caller is allowed to do
something, had to explicitly grant access for the root user with checking
for uid == 0. The new method exclusively grants access for the root user,
but specifies that others can possess the privilege as soon as they possess
the desktop-console privilege. This way, the lower lever application/script
does not have to handle any special case, it just has to check if the
caller (whether root or not) possesses the specific privilege.