fix fdi cache regen logic error
There's a logic error in the check of the return from fdi cache regen.
Basically if exit_type wasn't HALD_RUN_SUCCESS (which is defined as 0)
or return_code isn't 0 or 2, then an error occurred. Based on the
order users would have the following in their logs.

12:54:13.827 [I] mmap_cache.c:137: In regen_cache_cb exit_type=0, return_code=0
12:54:13.827 [E] mmap_cache.c:190: fdi cache regeneration failed!

The following patch corrects the issue and should be applied to 0.5.9
branch as well.