fix detection of libusb
Not sure why we check for libusb-config to find out if libusb and the
needed headers are available, but this fix this issue to the correct
check (e.g. for distributions where libusb and a libusb-devel
package exist). Fixed also NEWS file for correct libusb version.

Should also fix fd.o bug #8654