fix hald_runner_kill_all()
This patch fixes declaration of hald_runner_kill_all() which is declared
in the headerfile without argument, but in the code with a argument:
hald_runner_kill_all(HalDevice * device). Due to the usage of the function
in the code changed the code in hald/hald_runner.c to work without an
argument.

Fix also bugreport fd.o #8656