fix matching devices against dmi properties on the computer device
I've been tracing this, and it appears that when we do the match against
smbios.system.manufacturer then the following properties are on computer
are as follows:

 device udi = /org/freedesktop/Hal/devices/computer
  system.kernel.name = 'Linux'  (string)
  system.kernel.machine = 'i686'  (string)
  info.bus = 'unknown'  (string)
  info.udi = '/org/freedesktop/Hal/devices/computer'  (string)
  info.subsystem = 'unknown'  (string)
  info.product = 'Computer'  (string)
  system.kernel.version = '2.6.21-1.3194.fc7'  (string)

This leads me to think that we are not actually waiting for the dmi
prober to finish before we do the fdi scanning, and are thus not
matching.

1. We create computer with:

 device udi = /org/freedesktop/Hal/devices/computer
  system.kernel.name = 'Linux'  (string)
  system.kernel.machine = 'i686'  (string)
  info.bus = 'unknown'  (string)
  info.udi = '/org/freedesktop/Hal/devices/computer'  (string)
  info.subsystem = 'unknown'  (string)
  info.product = 'Computer'  (string)
  system.kernel.version = '2.6.21-1.3194.fc7'  (string)

2. We then add devices using coldplug_synthesize_events, doing
di_search_and_merge on each one as they are added.

3. We then run the probers and wait until they finish.

4. We then run di_search_and_merge on the computer device in
computer_probing_helper_done.

Now, the display matching quirks work because the device they are added
to is not a device added by coldplug (but computer itself), and thus
step 4 matches them and merges them. The laptop_panel quirks do not work
as the new device is added and di_search_and_merge'd before the probers
have finished, and we do not rescan devices doing absolute references
after step 4.

Let me do some more testing to verify, and see if I can come up with a
fix that doesn't break lots of other stuff.