[PATCH] add hald-runner support for singleton addons
Introduces a new method StartSingleton on org.freedesktop.HalRunner.  This
takes one parameter of an array of strings for the environment to launch the
singleton. hald-runner keeps a list of running singletons, which are killable
only by KillAll.

The patch also adds hald_runner_start_singleton for starting a singleton.