fix i8042 KBD keyboard mapping
We need to convert the large values displayed in dmesg to sensible values that
we can use with setkeycodes ioctl. We do this in the code rather than the fdi
as the value is displayed in dmesg, and we don't want to make it even more
complicated for the user.