don't compile Re-map keys without gperf
This fixes a problem with the "Re-map multimedia keys" feature while
compiling. If you have an older, or not up-to-date kernel the complete
compiling fail because hal-setup-keymap-hash-name.h may conatains
defines which would not work with your kernel version or your headers.

From now hal-setup-keymap-hash-name.h get always new genereated and
you are only able to compile the re-map feature if you have gperf
available.