add a refresh handler for linux devices
Some stuff in sysfs now sends out uevents when values change. A good
example of this is the power_supply class which sends out a uevent when
the percentage value changes or the battery or ac adapter is removed.

At the moment we ignore the uevents, but the attached patch adds a
refresh() method that can be used to update the values.