fix the input parameter for the SetPowerSave method
When we were fixing the uint/int retval problem a few months ago the
SetPowerSave method got changed from u SetPowerSave (b) to u
SetPowerSave (i). The hal-system-power-set-power-save script assumes the
bool rather than the signed int. This gives the error in
gnome-power-manager:

** WARNING **: ERROR:
 usage : hal-set-property --udi <udi> --key <key> aborting...

This also causes an abort if --g-fatal-warnings is used to run g-p-m.

I assume you just changed the b to an i as an oversight, in which case
the attached patch fixes the problem. If you do want to keep the script
input parameter as an int we'll have to fix up the shell we are using.