stop hald-addon-dell-backlight if needed kernel module isni't loaded
Currently the hald-addon-dell-backlight is running also if the
needed dcdbas kernel module is not loaded. In this case it provide
the brightness interface with only one brightness level, which is
useless.

Simply exit the addon if the kernel module is missing to prevent
to have a useless interface for brightness.