fix detection of system.formfactor=laptop for ACPI machines
There are problems on laptops if the smbios.chassis.type is not directly
mapped from the smbios to 'laptop' (e.g. smbios tell it's 'Unknown').

In this case HAL currently call as next coldplug_synthesize_events() after
the smbios probing and first then start to synthesizing the ACPI events/
devices. But in the ACPI code HAL set system.formfactor=laptop, if it isn't
already detected as Laptop (e.g. if the machine has a battery).

The problem is: all other devices as e.g. the WLAN device get synthesized
before the ACPI devices. Hence all fdi-file checks for laptop can fail with
the result:
 * the properties get not set as planed
 * e.g. the IPW killswitch device get not spawned

Fixed the problem for the ACPI case by check for a battery bay or a lid
button in /proc/acpi/*. we can ignore the APM case for the moment. Changed
acpi_synthesize() a little bit to let the function run without synthesize
ACPI devices because this function already check if the machine is a laptop.

In computer_probing_helper_done() HAL check now if system.formfactor isn't
set or not 'laptop' and calls acpi_check_is_laptop() in this case, to do the
needed checks and set if needed the machine as latop as soon as possible
before othere devices get handeled.