add new fdi directives *_outof
This introduce new fdi-file <match> directives for strings. The new are:
 - int_outof
 - string_outof
 - prefix_outof
 - contains_outof

They all work like the normal directives (string, prefix, contains), but only
on string properties and match against a ';'-separated list (any other ideas
for a char as separator?) as in this examples:

  <match key="usb.product_id" int_outof="0x4115;0x4116">
  - matches if "usb.product_id" is '0x4115' or '0x4115'

  <match key="system.hardware.product" string_outof="foo;bar">
  - matches if "system.hardware.product" is exact 'foo' or 'bar'

  <match key="system.hardware.product" prefix_outof="foo;bar">
  - matches if "system.hardware.product" starts with 'foo' or 'bar'

  <match key="system.hardware.product" contains_outof="foo;bar">
  - matches if "system.hardware.product" contains the (sub-)string
    'foo' or 'bar'

This new match directives allow us to reduce the size (and number of
duplicated rules) of many fdi files in hal and hal-info. As example I reduced:

  - 30-keymap-acer.fdi from 28532 to 8123 bytes (~ -72 %)
  - 10-usb-music-players.fdi from 78817 to 49860 Bytes (~ -37 %)