add soundcard device and fix device_file handling
Add the sound card device to HAL (e.g. /sys/devices/pci*/*/card0) to
represent the correct sysfs structure if HAL runs on a kernel which was
compiled with !CONFIG_SYSFS_DEPRECATED. The alsa and oss devices are
now childs of the releated sound card device.

Changed checks for empty device files path to check also for
device_file[0] == '\0'.