rework handling of data from udevinfo for more efficient memory usage
This patch reworks the handling of the output of udevinfo so that the output is
parsed into an intermediatary form (with a hash table of sysfpath to structures
pointing directly into the output as captured). This is converted into
HotplugEvents one at a time when needed by coldplug_get_hotplug_event.