process non-dependant hotplug events in parallel
Calculate event dependancy of hotplug events based of their sysfs_path. In
hotplug_event_process_queue if an event is dependant on an event in progress,
it will be held back in in the queue.

The event dependancy calculation is cribbed from udev. It requires the event
sequnce number in order to order the dependancies. We don't get event sequence
numbers during coldplug, so we emulate them using a simple counter.