use GString to convert a strlist propery to string instead of a static buffer
Instead of a static 256 chars buffer use a GString when converting strlist
properties to string. This prevents clipping the list.

Also fixes a hal crash because the old code didn't do a proper \0 terminate
when the string overflowed (strncpy is a very nice function *cough*).
Eventually the string containing garbage would be passed over dbus to the
runner which would disconnect from the private bus.