remove all childs of a device if it get removed
This change removes all child-devices of a device which get removed.

This solves problems e.g. with via fdi rules spawned devices, which currently
get not removed if their parent device get removed.

Currently if the parent get removed they are still there also if they are not
visible in lshal. This mean also, that all addons of a spawned device (as
e.g. the dell backlight addon) are still running. This patch should fix it.