prevent flood syslog if battery remaining time is above 60 hours
This prevents to get the syslog flooded with "remaining_time *very*
high, returning -1" warning messages on some laptops. With this patch the
message get send only the first 10 times and then only 1 message for each 100
cases (we can changes this also to a higher number).

We have this patch already since several months in the SUSE package, maybe
it's usefull in general.