fix endless loop in storage-addon
This fixes a endless loop if you attach e.g. a USB floppy device which
already contains a floppy media (because of the initial media_status). The
result is that the media in the device get never detected until the state
change again (remove/readd the media to the device).

IMO this could theoretically also happen on other devices which already
contains a media when attached.