ensure match rules are always stored before writing the jump_position in the cache
When a match rule is empty the expat end callback is called while the match
rule isn't saved yet. This patch ensures the match rule is saved _before_
saving the jump_positions, if this is not done then the jump_position will be
overwritten with 0.

Fixes hal aborting on Fujitsu laptops because the fujitsu fdi contains an
empty <match /> rule. Reported in
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=456314 and its  dupes